refactor: #40 - 랜딩페이지 문구 수정
