+++ src/features/landing/HeroSection.tsx
@@ -7,9 +7,9 @@ export function HeroSection() {
   return (
     <section className="relative overflow-hidden">
       {/* Gradient background */}
-      <div className="absolute inset-0 -z-10 bg-gradient-to-br from-amber-50 via-orange-50 to-rose-50 dark:from-amber-950/20 dark:via-orange-950/20 dark:to-rose-950/20" />
-      <div className="absolute -right-40 -top-40 -z-10 size-96 rounded-full bg-gradient-to-br from-amber-200/40 to-orange-200/40 blur-3xl dark:from-amber-800/10 dark:to-orange-800/10" />
-      <div className="absolute -bottom-20 -left-40 -z-10 size-80 rounded-full bg-gradient-to-tr from-rose-200/40 to-pink-200/40 blur-3xl dark:from-rose-800/10 dark:to-pink-800/10" />
+      <div className="absolute inset-0 -z-10 bg-linear-to-br from-amber-50 via-orange-50 to-rose-50 dark:from-amber-950/20 dark:via-orange-950/20 dark:to-rose-950/20" />
+      <div className="absolute -right-40 -top-40 -z-10 size-96 rounded-full bg-linear-to-br from-amber-200/40 to-orange-200/40 blur-3xl dark:from-amber-800/10 dark:to-orange-800/10" />
+      <div className="absolute -bottom-20 -left-40 -z-10 size-80 rounded-full bg-linear-to-tr from-rose-200/40 to-pink-200/40 blur-3xl dark:from-rose-800/10 dark:to-pink-800/10" />
 
       <div className="mx-auto max-w-5xl px-6 py-28 md:py-40">
         <h1 className="text-5xl font-bold tracking-tight text-center">
