+++ src/features/landing/LearningFlowSection.tsx
@@ -4,7 +4,7 @@ const scenes = [
     eyebrow: "기록",
     title: "적는 순간,\n복습이 설계됩니다",
     description:
-      "노트를 저장하면 1일, 3일, 7일 뒤 복습 일정이 자동으로 잡힙니다. 따로 캘린더를 쓰거나 알림을 설정할 필요가 없어요.",
+      "노트를 저장하는 순간, 복습 일정이 자동으로 잡힙니다. 캘린더를 따로 쓰거나 알림을 설정할 필요 없어요. 기록만 하면 딱다구리가 나머지를 챙깁니다.",
     mockup: <RecordMockup />,
     reverse: false,
   },
@@ -13,7 +13,7 @@ const scenes = [
     eyebrow: "알림",
     title: "잊어버리기 직전,\n딱 맞춰 알려드려요",
     description:
-      "에빙하우스 망각곡선에 따르면 기억은 학습 직후 가장 빠르게 사라집니다. 딱다구리는 기억이 흐릿해지는 바로 그 타이밍에 복습 알림을 보냅니다.",
+      "복습은 '해야지' 생각만으로는 잘 안 됩니다. 간격을 두고 반복할수록 기억이 오래 간다는 건 인지과학이 검증한 원리입니다. 딱다구리가 딱 그 타이밍에 먼저 찾아갑니다.",
     mockup: <NotificationMockup />,
     reverse: true,
   },
